requirements.txt
aniso8601==9.0.1
Authlib==1.0.1
boto3==1.23.9
botocore==1.26.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudevents==1.2.0
cryptography==37.0.2
decorator==5.1.1
deprecation==2.1.0
Flask==2.1.2
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
gunicorn==20.1.0
humanfriendly==10.0
idna==3.3
inuits-jwt-auth==0.0.26
inuits-otel-tracer==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
packaging==21.3
pika==1.2.1
Pillow==9.1.1
py==1.11.0
py-healthcheck==1.10.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.26
pytz==2022.1
rabbitmq-pika-flask==1.2.15
requests==2.27.1
retry==0.9.2
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
